+++ src/Color.hpp
@@ -21,10 +21,10 @@ namespace LCNS
         Color(void) = default;
 
         /// Constructor with the same value for each component
-        explicit Color(float commonValue);
+        explicit Color(double commonValue);
 
         /// Constructor with red, green and blue components (values between 0.0 and 1.0)
-        Color(float red, float green, float blue);
+        Color(double red, double green, double blue);
 
         /// Constructor with red, green and blue components (values between 0 and 255)
         Color(int red, int green, int blue);
@@ -39,10 +39,10 @@ namespace LCNS
         ~Color(void) = default;
 
         /// Component operator (read, write)
-        float& operator[](unsigned int index);
+        double& operator[](unsigned int index);
 
         /// Component operator (read only)
-        float operator[](unsigned int index) const;
+        double operator[](unsigned int index) const;
 
         /// Add 2 colors
         Color operator+(const Color& color) const;
@@ -57,10 +57,10 @@ namespace LCNS
         Color operator*=(const Color& color);
 
         /// Multiply a color by a scalar
-        Color operator*(float scale) const;
+        Color operator*(double scale) const;
 
         /// Scale a color
-        void operator*=(float scale);
+        void operator*=(double scale);
 
         /// Return true if all the coefficient for the 2 colors are the same, else return false
         bool operator==(const Color& color) const;
@@ -69,19 +69,19 @@ namespace LCNS
         bool isZero(void) const noexcept;
 
         /// Set the three component at the same time
-        void set(float red, float green, float blue) noexcept;
+        void set(double red, double green, double blue) noexcept;
 
         /// Set the three component at the same time
         void set(int red, int green, int blue) noexcept;
 
         /// Set red component (value between 0.0 and 1.0)
-        void red(float value) noexcept;
+        void red(double value) noexcept;
 
         /// Set green component (value between 0.0 and 1.0)
-        void green(float value) noexcept;
+        void green(double value) noexcept;
 
         /// Get blue component (value between 0.0 and 1.0)
-        void blue(float value) noexcept;
+        void blue(double value) noexcept;
 
         /// Set red component (values between 0 and 255)
         void red(int value) noexcept;
@@ -93,20 +93,20 @@ namespace LCNS
         void blue(int value) noexcept;
 
         /// Get red component (read only)
-        float red(void) noexcept;
+        double red(void) noexcept;
 
         /// Get green component (read only)
-        float green(void) noexcept;
+        double green(void) noexcept;
 
         /// Get blue component (read only)
-        float blue(void) noexcept;
+        double blue(void) noexcept;
 
     private:
         /// Utility method to set the components using parameters in the range [0, 255] as values in the range [0, 1]
         constexpr void _componentsIn0to1Range(int red, int green, int blue);
 
     private:
-        float _components[3] = { 0.0f, 0.0f, 0.0f };
+        double _components[3] = { 0.0, 0.0, 0.0 };
 
     };  // Class Color
